Implement multi processor compilation for dbuild targets.
Multi processor compilation is implemented with MultiToolTaskVisualD
which is a customized version of existing MultiToolTask from MSBuild.
The customization is needed to support 'One object file per package'
compilation model which cannot be supported with the original
MultiToolTask.

Limitations: The feature is not supported on VS2013 because the
underlying implementation (MultiToolTask etc.) does not exist in VS2013.
If the Multi-Processor flag is set in VS2013, it is simply ignored.