Fix for per package rebuild problem.
When compiling code with 'one object per package' option, the solution
would be fully recompiled each time.
The fix consists of two modifications:
1. Activate MaintainCompositeRootingMarkers (it is false by default in
the TrackedVCToolTask). This change correctly setups
CanonicalTrackedInputFiles and CanonicalTrackedOuputFiles when computing
the sources which need compilation.
2. Remove custom code which was used to save TLog files;
TrackedVCToolTask does this correctly in PostExecuteTool method. This
fixes the command line information saved in TLog files when compiling
per package.